data loss

+++ trunk/body-fs/src/com/bodyfs/ui/PrescriptionComposer.java
@@ -129,10 +129,11 @@ public class PrescriptionComposer extends GenericForwardComposer {
 		if (txtServices.getValue() != null && txtServices.getValue().length() > 0) {
 			saveServiceDate(txtServices.getValue(), prescription.getVisitDate(), prescription.getPersonId());
 		}
-		// Remove the patient from the queue
-		final IPersonDAO personDAO = (IPersonDAO) SpringUtil.getBean("personDAO");
-		personDAO.deleteQuickPatient(prescription.getPersonId());
-
+		if(event.getData().toString().equals("Done")){
+			// Remove the patient from the queue
+			final IPersonDAO personDAO = (IPersonDAO) SpringUtil.getBean("personDAO");
+			personDAO.deleteQuickPatient(prescription.getPersonId());
+		}
 		Clients.evalJavaScript("navigate('" + event.getData() + "')");
 	}
